fix warnings and error in gcc

+++ src/filter_sort_matrix.cpp
@@ -4,17 +4,14 @@
 using namespace Rcpp;
 
 // Custom comparison function for gene names (e.g., g1, g2, g10)
-bool geneCompare(const std::string &a, const std::string &b)
-{
+bool geneCompare(const std::string &a, const std::string &b) {
   size_t na = a.find_first_of("0123456789");
   size_t nb = b.find_first_of("0123456789");
 
-  if (na != std::string::npos && nb != std::string::npos)
-  {
+  if (na != std::string::npos && nb != std::string::npos) {
     std::string prefix_a = a.substr(0, na);
     std::string prefix_b = b.substr(0, nb);
-    if (prefix_a == prefix_b)
-    {
+    if (prefix_a == prefix_b) {
       return std::stoi(a.substr(na)) < std::stoi(b.substr(nb));
     }
   }
@@ -42,15 +39,13 @@ bool geneCompare(const std::string &a, const std::string &b)
 //'   targets = c("g3", "g4")
 //' )
 // [[Rcpp::export]]
-NumericMatrix filter_sort_matrix(NumericMatrix network_matrix,
-                                 Nullable<CharacterVector> regulators = R_NilValue,
-                                 Nullable<CharacterVector> targets = R_NilValue)
-{
+NumericMatrix
+filter_sort_matrix(NumericMatrix network_matrix,
+                   Nullable<CharacterVector> regulators = R_NilValue,
+                   Nullable<CharacterVector> targets = R_NilValue) {
   // Replace NA with 0
-  for (int i = 0; i < network_matrix.length(); i++)
-  {
-    if (R_IsNA(network_matrix[i]))
-    {
+  for (size_t i = 0; i < network_matrix.length(); i++) {
+    if (R_IsNA(network_matrix[i])) {
       network_matrix[i] = 0;
     }
   }
@@ -61,60 +56,47 @@ NumericMatrix filter_sort_matrix(NumericMatrix network_matrix,
 
   // Handle regulators filtering
   std::vector<std::string> filtered_regulators;
-  if (regulators.isNotNull())
-  {
+  if (regulators.isNotNull()) {
     CharacterVector reg(regulators);
     // Get intersection
-    for (int i = 0; i < curr_regulators.length(); i++)
-    {
+    for (size_t i = 0; i < curr_regulators.length(); i++) {
       std::string curr_reg = as<std::string>(curr_regulators[i]);
-      for (int j = 0; j < reg.length(); j++)
-      {
-        if (curr_reg == as<std::string>(reg[j]))
-        {
+      for (size_t j = 0; j < reg.length(); j++) {
+        if (curr_reg == as<std::string>(reg[j])) {
           filtered_regulators.push_back(curr_reg);
           break;
         }
       }
     }
-  }
-  else
-  {
-    for (int i = 0; i < curr_regulators.length(); i++)
-    {
+  } else {
+    for (size_t i = 0; i < curr_regulators.length(); i++) {
       filtered_regulators.push_back(as<std::string>(curr_regulators[i]));
     }
   }
 
   // Handle targets filtering
   std::vector<std::string> filtered_targets;
-  if (targets.isNotNull())
-  {
+  if (targets.isNotNull()) {
     CharacterVector tar(targets);
     // Get intersection
-    for (int i = 0; i < curr_targets.length(); i++)
-    {
+    for (size_t i = 0; i < curr_targets.length(); i++) {
       std::string curr_tar = as<std::string>(curr_targets[i]);
-      for (int j = 0; j < tar.length(); j++)
-      {
-        if (curr_tar == as<std::string>(tar[j]))
-        {
+      for (size_t j = 0; j < tar.length(); j++) {
+        if (curr_tar == as<std::string>(tar[j])) {
           filtered_targets.push_back(curr_tar);
           break;
         }
       }
     }
-  }
-  else
-  {
-    for (int i = 0; i < curr_targets.length(); i++)
-    {
+  } else {
+    for (size_t i = 0; i < curr_targets.length(); i++) {
       filtered_targets.push_back(as<std::string>(curr_targets[i]));
     }
   }
 
   // Sort gene names
-  std::sort(filtered_regulators.begin(), filtered_regulators.end(), geneCompare);
+  std::sort(filtered_regulators.begin(), filtered_regulators.end(),
+            geneCompare);
   std::sort(filtered_targets.begin(), filtered_targets.end(), geneCompare);
 
   // Create new matrix with filtered dimensions
@@ -124,20 +106,16 @@ NumericMatrix filter_sort_matrix(NumericMatrix network_matrix,
   std::unordered_map<std::string, int> old_reg_indices;
   std::unordered_map<std::string, int> old_tar_indices;
 
-  for (int i = 0; i < curr_regulators.length(); i++)
-  {
+  for (size_t i = 0; i < curr_regulators.length(); i++) {
     old_reg_indices[as<std::string>(curr_regulators[i])] = i;
   }
-  for (int i = 0; i < curr_targets.length(); i++)
-  {
+  for (size_t i = 0; i < curr_targets.length(); i++) {
     old_tar_indices[as<std::string>(curr_targets[i])] = i;
   }
 
   // Fill new matrix
-  for (int i = 0; i < filtered_regulators.size(); i++)
-  {
-    for (int j = 0; j < filtered_targets.size(); j++)
-    {
+  for (size_t i = 0; i < filtered_regulators.size(); i++) {
+    for (size_t j = 0; j < filtered_targets.size(); j++) {
       int old_row = old_reg_indices[filtered_regulators[i]];
       int old_col = old_tar_indices[filtered_targets[j]];
       result(i, j) = network_matrix(old_row, old_col);
@@ -148,12 +126,10 @@ NumericMatrix filter_sort_matrix(NumericMatrix network_matrix,
   CharacterVector new_regulators(filtered_regulators.size());
   CharacterVector new_targets(filtered_targets.size());
 
-  for (int i = 0; i < filtered_regulators.size(); i++)
-  {
+  for (size_t i = 0; i < filtered_regulators.size(); i++) {
     new_regulators[i] = filtered_regulators[i];
   }
-  for (int i = 0; i < filtered_targets.size(); i++)
-  {
+  for (size_t i = 0; i < filtered_targets.size(); i++) {
     new_targets[i] = filtered_targets[i];
   }
